requirements.txt
apex==0.9.10dev
classy_vision==0.6.0
facenet_pytorch==2.5.2
fuzzywuzzy==0.18.0
fvcore==0.1.5.post20220512
matplotlib==3.5.2
numpy==1.22.4
omegaconf==2.2.2
openpyxl==3.0.10
openpyxl_image_loader==1.0.5
pandas==1.4.2
Pillow==9.1.1
pysrt==1.1.2
PyYAML==6.0
scikit_learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
sentence_transformers==2.2.0
simcse==0.4
tensorboard==2.9.1
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
transformers==4.19.4
umap==0.1.1
vissl==0.1.6
xlsxwriter==3.0.3
